Finalize automated treatment of Facade subclasses Move facade related code to custom extension Remove inherited members from documentation Add show-inheritance option Add custom treatment of base classes
