+++ user_manual/source/API_2/API_model.rst
@@ -62,6 +62,6 @@ Detailed documentation
 .. automodule:: steps.API_2.model
     :members:
     :undoc-members:
-    :inherited-members: ndarray
     :special-members:
+    :show-inheritance:
