+++ user_manual/source/API_2/API_rng.rst
@@ -17,7 +17,7 @@ Detailed documentation
 .. automodule:: steps.API_2.rng
     :members:
     :undoc-members:
-    :inherited-members: ndarray
     :special-members:
+    :show-inheritance:
