+++ user_manual/source/API_2/API_saving.rst
@@ -24,7 +24,7 @@ Detailed documentation
 .. automodule:: steps.API_2.saving
     :members:
     :undoc-members:
-    :inherited-members: ndarray
     :special-members:
+    :show-inheritance:
