+++ user_manual/source/API_2/API_sim.rst
@@ -51,8 +51,8 @@ Detailed documentation
 .. automodule:: steps.API_2.sim
     :members:
     :undoc-members:
-    :inherited-members: ndarray
     :special-members:
+    :show-inheritance:
