+++ user_manual/source/API_2/API_utils.rst
@@ -26,24 +26,9 @@ Detailed documentation
 
 ----------
 
-.. autoclass:: steps.API_2.utils.Parameter
+.. automodule:: steps.API_2.utils
     :members:
     :undoc-members:
-    :inherited-members: ndarray
     :special-members:
+    :show-inheritance:
 
-.. autofunction:: steps.API_2.utils.ExportParameters
-
-.. autoclass:: steps.API_2.utils.NamedObject
-    :members:
-    :undoc-members:
-    :inherited-members: ndarray
-    :special-members:
-
-.. autoclass:: steps.API_2.utils.Params
-    :members:
-    :undoc-members:
-    :inherited-members: ndarray
-    :special-members:
-
-.. autofunction:: steps.API_2.utils.SetVerbosity
